Handle manually deleted Bitmap
If the user is rooted and goes in the data folder of the application and
delete a condition bitmap, the application will handle it this way:
 * If a detection is started with a missing bitmap, the condition associated
to this bitmap will always be considered as false.
 * In the click config dialog, the condition is displayed as a red cross
indicating the condition is corrupted.
 * In the click condition dialog, the error message is displayed, and the ok
button is grayed, leaving only the delete option to the user.