Use different positions for each orientation
Before this commit, the same position was saved for both orientation. This
lead to a bug where the overlay menu can be displayed outside the screen
bounds or to an unexpected visible position.

In order to fix this bug, the overlay menu saves a position for each
orientation, allowing to ensure the display within the screen bounds, as well
as consistency for the user.