Fix crash on condition verification
When trying to detect a condition with an area not contained in the screen, the detector was crashing. This was easily reproducible with a condition taken at the edge of the screen in one orientation, and then verifying in with another orientation.

This commit should fix this crash by simply checking if the area is contained in the screen size before any processing.