Fix overlay view drawing in system decors
It was not possible to draw the content of the overlay view on some part of the screen, such as the navigation bar, or the area around the notch.

This is now possible, as the overlay view is now drawn on the totality of the screen. This removes the complicated system of conversions between the view coordinates and the virtual display coordinates, as they now use the same system.

This should fix weird issues such as https://github.com/Nain57/Smart-AutoClicker/issues/12.

Note that it is impossible (or I haven't found a way) to draw OVER the navigation bar, so, even if the click or capture on it will be working, the selector will not be drawn (or partially). If the navigation bar is translucent, the overlay view will seen below.