Do not recreate VirtualDisplay on rotation
This is achieved by close the ImageReader and recreating a new one with the correct dimensions.
Then, update the VirtualDisplay surface by the new ImageReader one and finally, force the resize of the VirtualDisplay to the new display size.

This should workaround the new Android 14 limitation identified in https://github.com/Nain57/Smart-AutoClicker/issues/338 but this solution needs a bit of validation on previous Android versions before being considered stable