Robustify billing client connection
Client object seems to expire after a long time. As the use case of
a long living app like autoclickers is not really common, I don't think
Google took the time to QA it.

Anyway, a bug is opened in their issue tracker, let's hope it will be fixed
one day