Fix user consent state
Once the app is purchased and acknowledged, user consent is not necessary
as ads SDK is no longer enabled. Do not wait for consent upon app start in
those cases.

Previous condition was waiting for PURCHASED state, which is a transitional
state that is valid only during the purchase flow. We kown check for
PURCHASED_AND_ACKNOWLEDGED.