Remove useless screen bitmap copy after read
This was relevant with the old architecture, but isn't anymore as
the output image won't change anymore for the same ImageReader surface.

This was causing a lot of garbage collection and adding a lot of memory
pressure to the app, probably leading to high memory usage notification.

I have been looking for the root cause of those GC for long time and this
is finally the last piece of the puzzle, GC are now close to none during
detection